requirements.txt
PyYAML==5.1

